Rename modules to be shorter and more general
We remove the `_helper` suffix, as it doesn't provide any information.
We also remove the `grpc_` prefix for `streaming` as we have that info
in the class name, and it also involves Frequenz channels, and the same
for the `_strategy` in `retry`.

In general it is also a good practice to keep module names short.

Signed-off-by: Leandro Lucarella <luca-frequenz@llucax.com>