Add a `BaseApiClient` class
This class abstracts the common functionality of all the API clients.
It provides a way to connect and disconnect from the server, provided as
a URL, and a way to create the stubs to interact with the server.

It also provides an async context manager to make sure the connection is
closed when the client is done.

Signed-off-by: Leandro Lucarella <luca-frequenz@llucax.com>