Raise `ClientNotConnected` when getting the `stub` or `channel`
When the client is not connected there is no `stub` or `channel` that
can be retrieved, so we raise an exception instead of returning `None`.

This is much more convenient, as otherwise one must `assert is not None`
each time those properties are used, and most of the time one will use
them when it is expected that the client is connected.

Signed-off-by: Leandro Lucarella <luca-frequenz@llucax.com>