Add a function to call gRPC stubs and wrap errors
The new function `call_stub_method()` simplifies the common error
wrapping when calling stub methods, so errors are automatically
converted to `ApiClientError`s. It also checks that the client is
connected and raises an appropriate exception if it is not. Finally, if
a transform function is provided, the response will be automatically
transformed too.

Signed-off-by: Leandro Lucarella <luca-frequenz@llucax.com>