Use `netloc` from `urlparse` to specify the host
This is because `netloc` is formatted to be the network location, and
wraps ipv6 in square brackets, which we lose if we just use the `host`
field.

`netloc` also contains the port if a port was in the original uri.  We
need to append the default port only if the original uri didn't have a
port.

This fixes an issue when creating a grpc Channel with ipv6 addresses.

Signed-off-by: Sahas Subramanian <sahas.subramanian@proton.me>